requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
mypy==0.971
mypy-extensions==0.4.3
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.13
